Изменить FilmService
В FilmService:
-в методе getPopularFilms() теперь используется метод getPopular() из LikeDbStorage

+++ src/main/java/ru/yandex/practicum/filmorate/storage/LikeDbStorage.java
@@ -39,7 +39,7 @@ public class LikeDbStorage extends BaseRepository<Like> {
         updateRate(filmId);
     }
 
-    public List<Film> getPopular(int count) {
+    public List<Film> getPopular(Long count) {
         return jdbc.queryForList(GET_POPULAR_QUERY, Film.class, count);
     }
